# go_api_example/api/swagger.yaml
basePath: /
definitions:
  models.Collection:
    properties:
      content:
        type: string
      id:
        type: string
    type: object
info:
  contact:
    email: justine.bachelard@ext.uca.fr
    name: Justine Bachelard.
  description: API to manage collections.
  title: middleware/example
  version: 1.0.0
paths:
  /collections:
    get:
      description: Get collections.
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Collection'
            type: array
        "500":
          description: Something went wrong
      summary: Get collections.
      tags:
      - collections
    post:
      description: Create a new collection.
      parameters:
      - description: Content of the new collection
        in: body
        name: content
        required: true
        schema:
          type: string
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Collection'
        "400":
          description: Invalid request payload
        "500":
          description: Something went wrong
      summary: Create a new collection.
      tags:
      - collections
  /collections/{id}:
    get:
      description: Get a collection.
      parameters:
      - description: Collection UUID formatted ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Collection'
        "422":
          description: Cannot parse id
        "500":
          description: Something went wrong
      summary: Get a collection.
      tags:
      - collections
    put:
      description: Update an existing collection.
      parameters:
      - description: Collection UUID formatted ID
        in: path
        name: id
        required: true
        type: string
      - description: New content for the collection
        in: body
        name: content
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Collection'
        "400":
          description: Invalid request payload
        "404":
          description: Collection not found
        "500":
          description: Something went wrong
      summary: Update an existing collection.
      tags:
      - collections
schemes:
- http
swagger: "2.0"